refactor: Move ChainSyncClient, ChainSyncClientError, and to_traverse to new chain_sync_client module
